move `Database` to tests-integration
