+++ src/client/src/lib.rs
@@ -14,12 +14,10 @@
 
 mod client;
 pub mod client_manager;
-mod database;
 pub mod error;
 pub mod load_balance;
 mod metrics;
 pub mod region;
-mod stream_insert;
 
 pub use api;
 use api::v1::greptime_response::Response;
@@ -31,9 +29,7 @@ pub use common_recordbatch::{RecordBatches, SendableRecordBatchStream};
 use snafu::OptionExt;
 
 pub use self::client::Client;
-pub use self::database::Database;
 pub use self::error::{Error, Result};
-pub use self::stream_insert::StreamInserter;
 use crate::error::{IllegalDatabaseResponseSnafu, ServerSnafu};
 
 pub fn from_grpc_response(response: GreptimeResponse) -> Result<u32> {
