+++ tests-integration/src/lib.rs
@@ -15,12 +15,14 @@
 #![feature(assert_matches)]
 
 pub mod cluster;
+pub mod database;
 mod grpc;
 mod influxdb;
 mod instance;
 mod opentsdb;
 mod otlp;
 mod prom_store;
+mod stream_insert;
 pub mod test_util;
 
 pub mod standalone;
