+++ tests/runner/Cargo.toml
@@ -19,5 +19,6 @@ serde.workspace = true
 serde_json.workspace = true
 sqlness = { version = "0.5" }
 tempfile.workspace = true
+tests-integration.workspace = true
 tinytemplate = "1.2"
 tokio.workspace = true
